pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "refinery"
description = "Tools for refinement of sparse tracings into dense voxel-wise ground-truth."
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3"
]
dynamic = ["version", "readme"]

dependencies = [
    "numpy",
    "scipy",
    "zarr==2.17.2",
    "tifffile",
    "scikit-image",
    "toml",
    "dask==2024.4.1",
    "distributed==2024.4.1",
    "s3fs==2024.3.1",
    "boto3",
    "gcsfs==2024.3.1",
    "tensorstore",
    "jgo@git+https://github.com/scijava/jgo",
    "tqdm",
    "argschema",
    "seaborn",
    "pynrrd",
]

[project.optional-dependencies]
dev = [
    'black',
    'coverage',
    'flake8',
    'interrogate',
    'isort',
    'Sphinx',
    'pygit2'
]

[project.scripts]
transform = 'neuron_tracing_utils.transform:main'
align_to_cube = 'neuron_tracing_utils.align_to_cube:main'
fix_swcs = 'neuron_tracing_utils.fix_swcs:main'
refine = 'neuron_tracing_utils.refine:main'
astar = 'neuron_tracing_utils.astar:main'
fill = 'neuron_tracing_utils.fill:main'
render_mips = 'neuron_tracing_utils.render_mips:main'
resample = 'neuron_tracing_utils.resample:main'
stats = 'neuron_tracing_utils.stats:main'
organize_outputs = 'neuron_tracing_utils.organize_outputs:main'

[project.urls]
source = "https://github.com/AllenNeuralDynamics/neuron-tracing-utils"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "neuron_tracing_utils.__version__"}
readme = {file = ["README.md"]}

[tool.black]
line-length = 79
target_version = ['py39']
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
  )/
  | .gitignore
)
'''

[tool.coverage.run]
omit = ["*__init__*"]
source = ["neuron_tracing_utils", "tests"]

[tool.coverage.report]
exclude_lines = ["if __name__ == .__main__.:"]
fail_under = 0

[tool.isort]
line_length = 79

[tool.interrogate]
exclude = ["setup.py", "docs", "build"]
fail-under = 0